+++ packages/transaction-controller/src/utils/layer1-gas-fee-flow.test.ts
@@ -1,12 +1,12 @@
 import type { Provider } from '@metamask/network-controller';
 import type { Hex } from '@metamask/utils';
 
+import { updateTransactionLayer1GasFee } from './layer1-gas-fee-flow';
 import {
   TransactionStatus,
   type Layer1GasFeeFlow,
   type TransactionMeta,
 } from '../types';
-import { updateTransactionLayer1GasFee } from './layer1-gas-fee-flow';
 
 jest.mock('@metamask/controller-utils', () => ({
   ...jest.requireActual('@metamask/controller-utils'),
@@ -18,6 +18,7 @@ const LAYER1_GAS_FEE_VALUE_UNMATCH_MOCK: Hex = '0x2';
 
 /**
  * Creates a mock Layer1GasFeeFlow.
+ *
  * @param request - The request bag to create the mock
  * @param request.match - The value to return when calling matchesTransaction
  * @param request.layer1Fee - The value to return when calling getLayer1Fee